Renames ComputingElementBase to ComputingElement
